restructured directories and files, rendered Login component and added routing
