+++ dp-front-end/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 
-import Login from './components/Login';
+import Login from './components/login.js';
 import './App.css';
 
 
@@ -10,7 +10,8 @@ function App() {
   return (
     <Router>
       <div className="App">
-        <Route path="/login" component={Login} />
+        {/* <Route path="/login" component={Login} /> */}
+        <Login route path='/login' component={Login}/>
       </div>
     </Router>
   );
