--- dp-front-end/src/utils/axiosWithAuth.js
@@ -1,14 +1,0 @@
-import axios from 'axios';
-
-const axiosWithAuth = () => {
-    const token = localStorage.getItem('token');
-
-    return axios.create({
-        headers: {
-            'Content-Type': 'application/json',
-            'Authorization': `${token}`,
-        },
-    });
-};
-
-export default axiosWithAuth;
